Backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.0
django-cors-headers>=4.0
python-dotenv>=1.1
gunicorn>=20.1
psycopg2-binary>=2.9  # optional if using Postgres; can be removed for sqlite
requests>=2.25
